+++ project-core/src/main/java/com/spring/boot/bean/master/SysProject.java
@@ -13,7 +13,7 @@ public class SysProject {
     private Integer projectFinishedTotal;
     private Long projectEnergyId;
     private Date createTime;
-    private Integer companyId;
+    private Long companyId;
     private Integer year;
     private Integer month;
 
@@ -73,11 +73,11 @@ public class SysProject {
         this.createTime = createTime;
     }
 
-    public Integer getCompanyId() {
+    public Long getCompanyId() {
         return companyId;
     }
 
-    public void setCompanyId(Integer companyId) {
+    public void setCompanyId(Long companyId) {
         this.companyId = companyId;
     }
